// models/houseDetails.ts
export class HouseDetails {
    id: number;
    zpid: number;
    dateSold: Date;
    dateSoldString: string;
    description: string;
    isNonOwnerOccupied: boolean;
    lastSoldPrice: number
    mlsid: string
    monthlyHoaFee: string
    parcelId: string
    photoCount: number
    parentRegion: string;
    priceHistory: string;
    resoFacts: string;
    schools: string;
    streetAddress: string;
    taxAssessedValue: number;
    taxAssessedYear: number;
    taxHistory: string;
    yearBuilt: number;
}